Better HTTP get detection
Just checking if the $_POST array is empty is not enough. If you use verbs like PUT or DELETE, you do not want anything to be cached, since those operation would change the render. This change introduce a private function that will contain the logic regarding if the request is a GET (or HEAD) or not and the function is called both in intercept and in write.